Move hasAutometricsDecorator to utils
